Home Screen
This is a reference photo that we will use as an example, a goal for what the home screen should look like. The Atom logo should be a touchable item that opens up a menu for accessing other sections of the app. The individual blocks should be clicked on to show more information about a certain scouting session where additional information about the team is shown.